pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "novelmanga"
version = "0.1.0"
description = "Convert Aozora Bunko novels to manga pages automatically"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "google-genai>=1.0.0",
    "Pillow>=10.0.0",
]

[project.scripts]
novelmanga = "novelmanga.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/novelmanga"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.hatch.envs.default]
features = ["dev"]
